+++ test/utils/compatibility-test.js
@@ -16,6 +16,7 @@ vows.describe(Compatibility)
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isTrue(options.properties.merging);
         assert.isFalse(options.properties.spaceAfterClosingBrace);
+        assert.isFalse(options.properties.urlQuotes);
         assert.isTrue(options.units.rem);
         assert.isTrue(options.colors.opacity);
         assert.deepEqual(options.selectors.special, /(\-moz\-|\-ms\-|\-o\-|\-webkit\-|:dir\([a-z-]*\)|:first(?![a-z-])|:fullscreen|:left|:read-only|:read-write|:right)/);
@@ -61,6 +62,7 @@ vows.describe(Compatibility)
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isFalse(options.properties.merging);
         assert.isTrue(options.properties.spaceAfterClosingBrace);
+        assert.isFalse(options.properties.urlQuotes);
         assert.isTrue(options.properties.zeroUnits);
         assert.isFalse(options.units.rem);
         assert.isFalse(options.colors.opacity);
@@ -79,6 +81,7 @@ vows.describe(Compatibility)
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isFalse(options.properties.merging);
         assert.isTrue(options.properties.spaceAfterClosingBrace);
+        assert.isFalse(options.properties.urlQuotes);
         assert.isTrue(options.properties.zeroUnits);
         assert.isFalse(options.units.rem);
         assert.isFalse(options.colors.opacity);